+++ model/schema_group_test.go
@@ -1,0 +1,8 @@
+package model
+
+import "testing"
+
+func TestCount(t *testing.T) {
+	// TODO: Schema.Count
+	// TODO: Schema.CountAt
+}